requirements.txt
flask
flask-cors
torch
transformers
torchvision
Pillow
